I suggest improving the performance by using `bytes.Buffer`. Creating much `string` leads to such an overhead to both ns/op and allocations. This p/r also fixing too much json.Encoder for each map keys. Unlike #16, this p/r is safe in multiple goroutines. Note that this pr still depends on unsafe package but we can wait for https://github.com/fatih/color/issues/124 to be resolved. Or if you don't like this changes, you can close.
Performance difference is significant as follows.
```
benchmark              old ns/op     new ns/op     delta
BenchmarkFromat-16     12615         5766          -54.29%

benchmark              old allocs     new allocs     delta
BenchmarkFromat-16     191            50             -73.82%

benchmark              old bytes     new bytes     delta
BenchmarkFromat-16     6103          3033          -50.30%
```